+++ merkle_tree/tests/merkle_tree_hasher.rs
@@ -1,6 +1,4 @@
-use jf_merkle_tree::{
-    errors::MerkleTreeError, hasher::HasherMerkleTree, MerkleCommitment, MerkleTreeScheme,
-};
+use jf_merkle_tree::{errors::MerkleTreeError, hasher::HasherMerkleTree, MerkleTreeScheme};
 use sha2::Sha256;
 
 #[test]
@@ -10,9 +8,9 @@ fn doctest_example() -> Result<(), MerkleTreeError> {
     // payload type is `usize`, hash function is `Sha256`.
     let mt = HasherMerkleTree::<Sha256, usize>::from_elems(Some(2), my_data)?;
 
-    let root = mt.commitment().digest();
+    let commitment = mt.commitment();
     let (val, proof) = mt.lookup(2).expect_ok()?;
     assert_eq!(val, &3);
-    assert!(HasherMerkleTree::<Sha256, usize>::verify(root, 2, proof)?.is_ok());
+    assert!(HasherMerkleTree::<Sha256, usize>::verify(commitment, 2, val, proof)?.is_ok());
     Ok(())
 }